RAIN 6435 UI/UX Revamp:: Bill Genie Group Bill Search screen - Bill Download and Merge & Download integration (#2241)
* Implemented merge & download.

create api is behaving in two ways. Sometimes it is giving only one filestore id sometimes it is giving multiple ids ie., one id for every bill. So for now I have added download all bills function which opens every bill in a new tab. And if the create api returns only one filestore id then that is the desired result in that case url with the merged bills is opened twice.

* Update setupProxy.js

* Update index.js

* removed logs

* used mdms data instead of hardcoded keys